refactor: update backoff multiplier to use double type for retry configuration

+++ src/main/java/com/sprint/mission/discodeit/global/config/properties/StorageProperties.java
@@ -19,7 +19,7 @@ public record StorageProperties(
     public record Retry(
         @DefaultValue("3") @Min(1) int maxAttempts,
         @DefaultValue("2000") @Positive long backoffDelay,
-        @DefaultValue("3") @Positive int backoffMultiplier
+        @DefaultValue("3") @Positive double backoffMultiplier
     ) {
     }
 }
